+++ denops/translate/helper_test.ts
@@ -1,5 +1,5 @@
 import { assertEquals, Denops, test } from "./deps.ts";
-import { parseArgs } from "./util.ts";
+import { buildOption, defaultEndpoint } from "./helper.ts";
 
 const tests = [
   {
@@ -9,6 +9,8 @@ const tests = [
     bang: false,
     args: ``,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "en",
       target: "ja",
       text: "hello",
@@ -21,6 +23,8 @@ const tests = [
     bang: false,
     args: `"hello world"`,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "en",
       target: "ja",
       text: "hello world",
@@ -33,6 +37,8 @@ const tests = [
     bang: false,
     args: `ja en`,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "ja",
       target: "en",
       text: "hello\nworld",
@@ -45,6 +51,8 @@ const tests = [
     bang: false,
     args: `en ja "hello world"`,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "en",
       target: "ja",
       text: "hello world",
@@ -57,6 +65,8 @@ const tests = [
     bang: false,
     args: `en ja "hello world" gorilla`,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "en",
       target: "ja",
       text: "hello world gorilla",
@@ -69,6 +79,8 @@ const tests = [
     bang: false,
     args: `en ja "hello world" "gorilla"`,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "en",
       target: "ja",
       text: "hello world gorilla",
@@ -81,6 +93,8 @@ const tests = [
     bang: false,
     args: `en ja "hello 'world' "gorilla"`,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "en",
       target: "ja",
       text: "hello 'world' gorilla",
@@ -93,6 +107,8 @@ const tests = [
     bang: true,
     args: `"hello world"`,
     want: {
+      endpoint: defaultEndpoint,
+      isDeepL: false,
       source: "ja",
       target: "en",
       text: "hello world",
@@ -108,7 +124,7 @@ for (const tt of tests) {
       if (tt.name === "2 args" || tt.name === "no args") {
         await denops.call("setline", 1, ["hello", "world"]);
       }
-      const got = await parseArgs(denops, tt.bang, tt.start, tt.end, tt.args);
+      const got = await buildOption(denops, tt.bang, tt.start, tt.end, tt.args);
       assertEquals(got, tt.want);
     },
   });
